refactor: useResponsiveImages receives an array of `ImageObj`
